use typed seed export deep link
